Fix inconsistent state that happened when spawning a blip inside an activatable block.
The bug happened when a blip moved into that block.
Then, the blip is removed after activation. However,
the blip inside the block continues living, since
blocks are activated only by *moving* blips. When the
moving blip is removed, the blip state is set to None
at that position, causing the grid to be inconsistent
with the vector.